feat: implement WorkerBinding as an enum
Currently the `WorkersBinding` assume that it always have `namespace_id` and also does not take into consideration all the other possible fields of different bindings.
This changes so that the `WorkerBinding` is now an enum with all the possible options listed on https://developers.cloudflare.com/workers/configuration/multipart-upload-metadata/#bindings.